Cleanup

+++ src/QACG/CircGen/Add/SimpleRipple.hs
@@ -18,7 +18,7 @@ import QACG.CircGen.Bit.Toffoli
 -- |Generates the addition circuit in <http://arxiv.org/abs/quant-ph/0410184> and returns it as a circuit
 mkSimpleRipple :: [String] -> [String] -> String -> Circuit
 mkSimpleRipple aLns bLns carry = circ
-  where (_,(_,_,circ)) = runState go ([], ['c':show x|x<-[0..10]] , Circuit (LineInfo [] [] [] []) [] [])
+  where (_,(_,_,circ)) = runState go ([], ['c':show x|x<-[0::Int .. 10]] , Circuit (LineInfo [] [] [] []) [] [])
         go             = do (aOut,bOut) <- simpleRipple aLns bLns carry
                             _ <- initLines aLns
                             _ <- initLines bLns
@@ -27,7 +27,7 @@ mkSimpleRipple aLns bLns carry = circ
 
 mkSimpleSubtract :: [String] -> [String] -> Circuit
 mkSimpleSubtract aLns bLns = circ
-  where (_,(_,_,circ)) = runState go ([], ['c':show x|x<-[0..10]] , Circuit (LineInfo [] [] [] []) [] [])
+  where (_,(_,_,circ)) = runState go ([], ['c':show x|x<-[0::Int .. 10]] , Circuit (LineInfo [] [] [] []) [] [])
         go             = do (aOut,bOut) <- simpleSubtract aLns bLns 
                             _ <- initLines aLns
                             _ <- initLines bLns
@@ -36,7 +36,7 @@ mkSimpleSubtract aLns bLns = circ
 
 mkSimpleCtrlRipple :: String -> [String] -> [String] -> String -> Circuit
 mkSimpleCtrlRipple ctrl aLns bLns carry = circ
-  where (_,(_,_,circ)) = runState go ([], ['c':show x|x<-[0..10]] , Circuit (LineInfo [] [] [] []) [] [])
+  where (_,(_,_,circ)) = runState go ([], ['c':show x|x<-[0::Int .. 10]] , Circuit (LineInfo [] [] [] []) [] [])
         go             = do (aOut,bOut) <- simpleCtrlRipple ctrl aLns bLns carry
                             _ <- initLines aLns
                             _ <- initLines bLns
